+++ index.js
@@ -1,6 +1,6 @@
-import { DateTime } from './modules/luxon/src/luxon.js';
 import { btnsClick, listSection, hideSection } from './modules/eventlisteners.js';
 import { BookList } from './modules/bookclass.js';
+import { DateTime1 } from './modules/datetime1.js';
 
 btnsClick();
 
@@ -10,16 +10,7 @@ const book = new BookList();
 btnAdd.addEventListener('click', () => {
   book.addBook();
 });
-const DateTime1 = () => {
-  const datetimeOutput = document.getElementById('datetime-output');
 
-  const updateDateTime = () => {
-    const now = DateTime.now().toLocaleString(DateTime.DATETIME_MED_WITH_SECONDS);
-    datetimeOutput.textContent = now;
-  };
-  updateDateTime();
-  setInterval(updateDateTime, 1000);
-};
 const loadPage = () => {
   hideSection();
   listSection.classList.remove('hidden');
